requirements.txt
asgiref==3.8.1
Django==4.2.16
django-etc==1.4.0
django-hitcount==1.3.5
django-modeltranslation==0.19.11
pillow==11.0.0
python-decouple==3.8
sqlparse==0.5.2
typing_extensions==4.12.2
whitenoise==6.8.2